Refactor content service file
Refactor src/services/airtable/contentService.ts into smaller, more focused files.
